requirements.txt
# Interface Gráfica Moderna
customtkinter>=5.2.0

# Manipulação de Imagens e Leitura de EXIF (Metadados)
Pillow>=10.0.0

# Ferramentas para criar o executável (.exe)
pyinstaller>=6.0.0
auto-py-to-exe>=2.40.0